Moves three views into class as children, buttons out. Displays
The selector buttons are a separate component which is passed the
click functions on the prop object. Note that the name _must_
start with a capital letter (wasn't previously).

To get the this binding to work, moved the view type components
into the class. Three buttons (which do nothing) are now displayed.