Buttons change state, new component uses state to render correct view
I was confused about the relationship of the buttons to the view --
I had been trying to use the buttons to swap in a new component,
which is a pretty Angular way of doing things. Instead, the buttons
simply change the ImageGallery app's "view" state, and then a
second component uses that state to select which component to display.

Successfully changes view components in and out.