Moves image html creation into a map to render all images on props
Instead of rendering just a single image, the GalleryView now takes
in an array of images and maps over them, generating the img tag
and html for each image in the array, returning the completed list
of images (with unique 'keys' for the divs containing the images).

Correctly displays all 4 images.