Removes render from App, index is only render - to pass jest tests
Jest snapshot tests were failing because the render method was being
called in the same module as the app, but the tests don't have access
to a DOM to mount components into.

This copies the old index into App and removes just the render function,
and then removes all of the other scripts from the index.js

Works.