Adds AddImage component connections to App, moves images into state
The AddImage component does all of the work and logic for the form
of getting a new image together, it then passes it to the addImage
method on the ImageGallery component that updates the state of the
image array. Because the array is in the state, any changes to it
automatically cause a re-rendering of the selected component.

Also cleans up imports.